backend/requirements.txt
Django==4.2.30
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2-binary>=2.9
dj-database-url==2.3.0
django-storages[s3]==1.14.4
boto3==1.35.81
Pillow==10.4.0
python-dotenv==1.0.1
sib-api-v3-sdk==7.6.0
gunicorn==22.0.0
whitenoise==6.8.2
django-filter==24.3
